pyproject.toml
[project]
name = "smarthealth-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.12.0",
    "gravixlayer>=0.0.30",
    "httpx>=0.28.1",
    "langchain-community>=0.3.29",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.33",
    "loadenv>=0.1.1",
    "openai>=1.108.1",
    "pip>=25.2",
    "pybind11<3.0",
    "pydantic>=2.11.9",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "setuptools>=80.9.0",
    "spacy>=3.8.7",
    "streamlit>=1.49.1",
    "wheel>=0.45.1",
]
